+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # CHANGELOG
 
+## next release
+
+- Add command to reveal the current project in the file explorer (#130)
+- Rename all commands to be sentence-case, matching rest of Obsidian
+- Support templates when creating scenes (Templater, core Templates plugin) (#134)
+
+## 2.0.3
+
+- Correct bug in previous set of 2.0.2 fixes.
+
+## 2.0.2
+
+- Use new atomic frontmatter APIs (#136, #133)
+- Slightly better fallbacks for null project states (#132)
+
+## 2.0.1
+
+- Add funding URL
+- Fix #123
+
 ## 2.0.0
 
 Total rewrite.